WIP implementation of "sets of projective numbers."
I realized that I need this for testing operations on bounds. The great
thing about Sopn's is that they don't really have any edge cases. It's
trivial to lift arithmetic over Pnums to arithmetic over Sopns.

The implementation here is kind of crappy. It will look a lot better
when I define iterators for bounds and Sopns.