Implement exp.
Taking an "easy way out" approach of using floating point exp, and
widening if it lands on exact value when the input is not 0 (since)
exp of a rational should never be a rational. This could return a
3-ulp wide Pbound when we could have returned a 1-ulp wide Pnum, but
I'm not so concerned about it for a prototype, and this approach is
better than returning an exact value inappropriately.