Change Pbound packed storage format.
To align the Pnums on sizeof(Pbound)/2 boundaries. Doesn't matter for
current single-byte Pbounds, but the theory is that this will be better
for larger types. I'm assuming there is probably some efficient way to
get the upper and lower halves of larger integer types.

However, I'm starting to question the utility of storing the "empty"
tag as part of the packed Pbound. I think it might end up being better
to have packed non-empty Pbounds, and then wrap them in a type that
stores emptyness as a bool.

In practice, if we're storing a bunch of Pbounds, it will almost
certainly be because we want to represent a union of disjoint Pbounds.
There's no point in putting empty Pbounds in such a union, so we could
pack known non-empty Pbounds into the union type, and then pull them
back out to the safer maybe-empty type when we read them.

Actually, the more I write about this, the better it seems. It makes a
lot of sense to define arithmetic over non-empty Pbound inputs (but
still with maybe-empty pbound outputs). This would let us separate
sepcial case empty logic from all the other logic.

Finally, it would be great to not have to use weird-sized Pbounds to
allow easily stealing a byte from them.