Change encoding of Pbounds.
Pbounds are now a composite type with an isempty Bool, and a packed
NonEmptyPbound. This means we don't have to steal any bits from Pnums
to put them in a NonEmptyPbound.

I think that most of the times you
might want to store a large group of Pbounds, it is probably because
you want to represent a disjoint union. We can make a special type for
this that internally stores NonEmptyPbounds, and therefore doesn't pay
the cost of the extra "isempty" Bool.